% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dostata-package.R
\docType{package}
\name{dostata-package}
\alias{dostata}
\alias{dostata-package}
\title{dostata: Run Stata Commands from R}
\description{
An R to Stata interface that allows Stata commands to be executed from within R. Data sets can easily be transferred between R and Stata, and Stata estimation results can be returned to R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/james-atkins/dostata}
  \item Report bugs at \url{https://github.com/james-atkins/dostata/issues}
}

}
\author{
\strong{Maintainer}: James Atkins \email{james@jamesatkins.com}

}
\keyword{internal}
