% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dostata.R
\name{stata_run}
\alias{stata_run}
\title{Run a Stata command}
\usage{
stata_run(
  commands,
  ...,
  session = stata_default_session(),
  env = parent.frame()
)
}
\arguments{
\item{commands}{A vector of Stata commands to run}

\item{...}{These dots are for future extensions and must be empty.}

\item{session}{The Stata session to run the commands in}

\item{env}{The R environment in which to look for data frames to transfer to Stata}
}
\description{
Run a Stata command
}
