% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dostata.R
\name{stata_start_session}
\alias{stata_start_session}
\title{Start a new session of Stata running in batch mode}
\usage{
stata_start_session(path = stata_path())
}
\arguments{
\item{path}{The path to the Stata executable}
}
\description{
Start a new session of Stata running in batch mode
}
